Safari-609.1.11.1/LayoutTests/http/tests/cache/disk-cache/resource-becomes-uncacheable-expected.txt

Tests that resources are removed from the cache if they become uncacheable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Resource should be in the cache now.
Load resource again using default cache policy.
PASS internals.xhrResponseSource(xhr) is "Disk cache"

Now load again the same resource, ignoring the cached data.
This time the resource will not be cacheable
PASS internals.xhrResponseSource(xhr) is "Network"

Stale resource should have been removed from the cache.

Now try to load the resource from the cache.
PASS internals.xhrResponseSource(xhr) is "Network"
PASS successfullyParsed is true

TEST COMPLETE
